fix: sdk move player avatars lerping (#1397)
WHAT?
Added an immediate bool to the position package shared by the comms peers. When the AvatarSystem updates a fellow avatar position, if that immediate bool is true, we remove the entity from the engine, update its transform position, and then add it back to the engine. That way we avoid all the trouble around adding a new property in the transform that would only be used for this case.