fix: hover feedback when blocked by ui (#1563)
This PR adds an extra graphics ray cast (only UI, and only when the pointer event interaction ray already detected an interaction) to detect if there is UI in the middle of the screen, where the pointer is. In that case, we consider the UI as blocker to the interaction of the object behind it.