Horizon mask. Rotation of textures. Different speed for each axis. New test with satellite layer
Added a texture to mask the horizon, with a values to control the size and the offset. Added a value to rotate the textures of the layers. Now the panning speed of the layers are in separated X and Y values, and are no longer multiplied with the offset. New test for the satellite layers in order to get better results when rotating.