refactor: change SceneMetricsController approach to use DataStore for rendereables and meshes.
- had to change the general approach because the meshes were already read-only when assets were reused. This means that triangle counting
only worked the first time the asset was loaded, and couldn't be re-calculated after that. To fix this issue without degrading the code,
moved the triangles count down to the poolable Assets on the APK level.
LoadWrapper_GLTF and ParametrizableShape bring the APK triangle count to the DataStore, and the SceneMetricsController use the DataStore to
 calculate the final metrics.