Mask below horizon. Plane color below the horizon. Reflection probe added.
Added a mask for the layers that mask everything below the horizon. Also added a plane projected on that mask. Added a color for the plane and it's opacity and a float value to control the height of the mask and the plane color. Added reflection probe to test reflections baking.