feat: receive improved friend requests (#3721)
Take in consideration that all the information displayed in zone is fake.

Go to: .zone
Open the friend request list, you should see faked friend requests
Click any of the received requests
Reject the request. It should be removed from the list
Open another received request
Accept it
Click on any of the sent requests
Cancel it. It should be removed from the list
Go to: .org
The friend requests should be displayed normally and all the interactions should be working as usual