Fix: E&F action keys overlaps with chat (#5268)
## How to test the changes?

1. Launch the explorer (Web, as a guest)
2. Go to Antrom RPG scene (`position=143,-10`) -> You will see UI about "New Player Tutorial" with two buttons - for E and F shortcuts
3. Click on the free space on the screen, so the cursor get permanently locked
4. Open chat by pressing Enter on the keyboard
5. Type E in the chat - as a result "New Player Tutorial" UI should not be closed by this 

I suggest to check this test scenario on the `prod` to verify reproducibility of the test steps for the bug itself

## What does this PR change?
Fix #4871 

Added additional check in the `PointerEventsController.cs` to not process events if chat is opened. This is similar to already existing check there (lines 245-251), but without relying on the `chatCamera`

## Copilot summary
Refactored legacy input controller code and added a condition to avoid pointer events when chat input is visible. The changes affect the `InputController_Legacy` and `PointerEventsController` classes in the `UUIDEventComponentsPlugin` folder. The goal is to improve the code quality and the user experience of the plugin.