Chore: updating User Profile with null model or null avatar  (#5350)
## How to test the changes?

Since it was an edge-case, I am not sure it can be tested by QA. 
However, I expect it to fix issue #5058 (if it still present). In case it is reproducible, you should see a guy in just black pants with normal skin color (instead of black in world and transparent in AvatarEditor)

## What does this PR change?
Closes #4755

- add fallback or recovering UserProfile in the case it was failing from payload parsing

## Copilot summary

This pull request enhances the `UserProfile` feature by using new C# features, adding error handling, and removing unused code. It also updates the `ProfileHUDViewV2` and `AvatarModel` classes to use new C# syntax for readability and consistency. Finally, it adds a new assembly reference to the `UserProfile.asmdef` file to enable feature flagging.